% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ciHS_demographics.R
\name{ciHS_demographics}
\alias{ciHS_demographics}
\title{Determine the demographic breakout of College In The High School students by High School}
\usage{
ciHS_demographics(data, high_school_name = "")
}
\arguments{
\item{data}{The student table saved as an object}

\item{high_school_name}{You can write out the entire name of the high school or match based on given name.}
}
\value{
A tibble will be returned
}
\description{
Determine the demographic breakout of College In The High School students by High School
}
\examples{

ciHS_demographics(student_tbl)


student_tbl \%>\%
 ciHS_demographics(high_school_name = "Eph") \%>\%
 filter(year >= "B23")

}
