% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_dw_transcript.R
\name{clean_dw_transcript}
\alias{clean_dw_transcript}
\title{Clean DW Transcript Table}
\usage{
clean_dw_transcript(data)
}
\arguments{
\item{data}{name of object where the transcript data is stored from data warehouse}
}
\description{
Used to clean transcript table by removing non-credit course, community courses, or any other not applicable.
See .R file for specific details
}
\examples{


}
