% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_ids.R
\name{clean_ids}
\alias{clean_ids}
\title{Clean IDs}
\usage{
clean_ids(data, check_errors = FALSE)
}
\arguments{
\item{data}{Any dataframe or list object contains sids}

\item{check_errors}{Returns a column indicating whether there are errors for a given sid or emplid. Errors
are deemed as not having the right character length (9) or containing lettters in them.}
}
\value{
Returns a dataframe or list with clean sids.
}
\description{
Use this function to remove SIDs or Emplids that may have extra spaces or hyphens in them.
}
\examples{
library(magrittr)
library(dplyr)

messy_sids <- tibble(
  SID = c("123456789", "123-456-789", " 1234-56789", "12345"),
  Emplid = c("777777777", "8888-8989", " d34df8878", "77777"),
  name = c("Batman", "Supergirl", "Flash", "Aquaman")
)

clean_ids(messy_sids)

messy_sids \%>\%
  clean_ids(check_errors = TRUE)


}
