% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_sid.R
\name{clean_sids}
\alias{clean_sids}
\title{Clean SIDs}
\usage{
clean_sids(data, check_length = FALSE)
}
\arguments{
\item{data}{Any dataframe or list object contains sids}

\item{check_length}{Return a column that indicates whether an sid is the correct lenght, 9 characters.}
}
\value{
Returns a dataframe or list with clean sids.
}
\description{
Use this function to remove SIDs that may have extra spaces or hyphens in them.
}
\examples{
library(magrittr)
library(dplyr)

messy_sids <- tibble(
  SID = c("123456789", "123-456-789", " 1234-56789", "12345"),
  name = c("Batman", "Supergirl", "Flash", "Aquaman")
  )

clean_sids(messy_sids)

messy_sids \%>\%
  clean_sids()


}
