% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/course_success_rate.R
\name{course_success_rates_publ}
\alias{course_success_rates_publ}
\title{Publication ready course success rates}
\usage{
course_success_rates_publ(academic_year)
}
\arguments{
\item{academic_year}{Designate academic year using codes.}
}
\value{
A tibble containing course success rates in a summarized fashion.
}
\description{
Publication ready course success rates
}
\examples{
\dontrun{

course_success_rates_publ("C01")

course_success_rates_publ("B56")

}

}
