% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/race_ethnic_trans.R
\name{race_ethnic_trans}
\alias{race_ethnic_trans}
\title{Allows the user to convert numberical codes for racial/ethnic codes into A/W or HUG grouping for BBCC.}
\usage{
race_ethnic_trans(data)
}
\arguments{
\item{data}{A dataframe containing desired data}
}
\value{
A new column containing recoded data
}
\description{
Allows the user to convert numberical codes for racial/ethnic codes into A/W or HUG grouping for BBCC.
}
\examples{

df \%>\%
  race_ethnic_trans()


}
