% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retention.R
\name{retention}
\alias{retention}
\title{Student Rention}
\usage{
retention(data, cohort_yr, nxt_yr)
}
\arguments{
\item{data}{A R object with the STUDENT table}

\item{cohort_yr}{The academic year of the starting cohort}

\item{nxt_yrq}{The following academic year after the cohort}
}
\value{
A data frame with demographics and retention for each student in the designated
starting cohort
}
\description{
The resulting data set contains columns with information about a student
returning in subsquent quarters after their initial quater.
}
\examples{
\dontrun{

student \%>\%
  retention('B90', 'C01')

}



}
\concept{return}
