% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retention.R
\name{retention_summarized}
\alias{retention_summarized}
\title{Summarized Student Retention}
\usage{
retention_summarized(data, desired_qtr)
}
\arguments{
\item{data}{A data set containing retention data}

\item{desired_qtr}{The quarter for which you want to get retention data for}
}
\value{

}
\description{
The resulting data set is a summarized view of rention data.
}
\examples{

\dontrun{

retention_data \%>\%
  retention_summarized(winter)



student \%>\%
  retention('B90', 'C01') \%>\%
  retention_summarized(winter)

}

}
\concept{retention}
