% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_demographics.R
\name{rs_demographics}
\alias{rs_demographics}
\title{Determine the demographic breakout of Running Start Students by High School}
\usage{
rs_demographics(data, high_school_name = "")
}
\arguments{
\item{data}{The student table saved as an object}

\item{high_school_name}{You can write out the entire name of the high school or match based on given name.}
}
\value{
Returns a tibble where ethnic groups are grouped according to typical categories (A/W, HUG) for each year and high school selected.
}
\description{
Determine the demographic breakout of Running Start Students by High School
}
\examples{


 rs_demographics(high_school_name = 'Eph')

 student_tbl \%>\%
    rs_demographic() \%>\%
    filter(year == "B78")

}
