% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_fte.R
\name{state_fte}
\alias{state_fte}
\title{state_fte}
\usage{
state_fte(data, yrq)
}
\arguments{
\item{data}{Will be the dbo class table saved as an R object}

\item{yrq}{The desired year quarter.}
}
\value{
returns a tibble.
}
\description{
state_fte
}
