% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stu_qtr_enrollment.R
\name{stu_not_registered_nxt_qtr}
\alias{stu_not_registered_nxt_qtr}
\title{Check student enrollment in a quarter}
\usage{
stu_not_registered_nxt_qtr(current_qtr, summer_qtr = FALSE)
}
\arguments{
\item{current_qtr}{This is the last quarter in which a student attended.}

\item{summer_qtr}{Providing a field will allow for the inclusion of a summer quarter.}
}
\value{
A tibble displaying which students returned in the designated quarters.
}
\description{
Running this function will allow one to find the last quarter in which a student is registered in.
}
\examples{
\dontrun{

stu_not_registered_nxt_qtr("B902", summer = TRUE)

}

}
