% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stu_qtr_enrollment.R
\name{stu_not_registred_nxt_qtr}
\alias{stu_not_registred_nxt_qtr}
\title{Student Quarter Enrollment}
\usage{
stu_not_registred_nxt_qtr(current_qtr, nxt_qtr)
}
\arguments{
\item{current_qtr}{This is the quarter of which you want to start comparing.}

\item{nxt_qtr}{This is the quarter at which you want to see if a student is enrolled in.}
}
\description{
Running this function will allow one to see if a student is registered in designated quarter given the current quarter.
}
\examples{

stu_not_registred_nxt_qtr("C012", "C013")



}
