% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ciHS_demographics.R
\name{ciHS_demographics}
\alias{ciHS_demographics}
\title{Determine the demographic breakout of CiHs by High School}
\usage{
ciHS_demographics(data, high_school_name = "")
}
\arguments{
\item{data}{The DW STUDENT table saved as an R object.}

\item{high_school_name}{Provide the entire or partial High School name.}
}
\value{
A tibble with number of students per high school broken out by race/ethnicity.
}
\description{
Determine the demographic breakout of CiHs by High School
}
\examples{
library(tibble)
library(magrittr)

# The data will look something like this.
df <- tibble(
   sid = c("1" , "2"),
   year = c("B90", "B90"),
   hi_schl = c("178", "179"),
   race_ethnic_code = c("1", "6"),
   dual_enroll = c("2", "1")
   )

 df


# Providing no high school will return all the values
 df \%>\%
    ciHS_demographics()


# Using either the full name or paritial name will suffice if spelled correctly.
 df \%>\%
    ciHS_demographics(high_school_name = "Moses Lake")


}
