% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_dw_transcript.R
\name{clean_dw_transcript}
\alias{clean_dw_transcript}
\title{Clean DW Transcript Table}
\usage{
clean_dw_transcript(data)
}
\arguments{
\item{data}{name of object where the transcript data is stored from data warehouse}
}
\value{
A dataframe with a cleaned transcript table
}
\description{
Used to clean the transcript table by removing non-credit course, community courses, or any other not applicable.
}
\examples{
transcript_tbl \%>\%
  clean_dw_transcript()



}
