% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_sid.R
\name{clean_sids}
\alias{clean_sids}
\title{Clean SIDs}
\usage{
clean_sids(data)
}
\arguments{
\item{data}{Any dataframe or list object contains sids}
}
\value{
Returns a dataframe or list with clean sids.
}
\description{
Use this function to remove SIDs that may have extra spaces or hyphens in them.
}
\examples{
library(magrittr)
library(dplyr)

messy_sid <- c("1234", "123-4", " 1234-5")

clean_sids(messy_sid)


tibble(messy_sid) \%>\%
   mutate(sid = clean_sids(messy_sid))


}
