% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_sid.R
\name{clean_sids}
\alias{clean_sids}
\title{Clean SIDs}
\usage{
clean_sids(data)
}
\arguments{
\item{data}{Any dataframe object containing a column named 'sid'}
}
\value{
A dataframe containing a new column, 'clean_sid', with cleaned sid numbers.
}
\description{
Use this function to remove SIDs that may have extra spaces or hyphens in them.
}
