% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/course_success_rate.R
\name{course_success_rate}
\alias{course_success_rate}
\title{Calculating course success rates}
\usage{
course_success_rate(data)
}
\arguments{
\item{data}{}
}
\description{
Calculating course success rates
}
\examples{


}
