% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/course_success_rate.R
\name{course_success_rate}
\alias{course_success_rate}
\title{Calculating course success rates}
\usage{
course_success_rate(data, rate = "raw")
}
\arguments{
\item{rate}{Determine the output type. Default is the raw output but can be changed to 'percent' to get
pass rate as a percentage}
}
\value{
A dataframe of all the key metrics for course success rates
}
\description{
Calculating course success rates
}
\examples{
\\donotrun{
transcript_tbl \%>\%
   course_success_rates()

transcript_tbl \%>\%
   course_success_rates(rate = 'percent')

}


}
