% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_fte_contribution.R
\name{rs_fte_contribution}
\alias{rs_fte_contribution}
\title{Calculate the amount of ftes that running start students contribute to the colleges overall fte count for each quarter.}
\usage{
rs_fte_contribution(data, no_summer = FALSE)
}
\arguments{
\item{data}{A dataframe of the student table}

\item{no_summer}{Determine whether to exclude summer quarter as part of the output. By default it is set to FALSE}
}
\value{
A tibble will be returned
}
\description{
Calculate the amount of ftes that running start students contribute to the colleges overall fte count for each quarter.
}
\examples{

student \%>\%
  rs_fte_contribution()

student \%>\%
  rs_fte_contribution() \%>\%
  filter(str_detect(yrq, "B[6-9][0-9][2-4]"))

rs_fte_contribution(student)


}
