% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_fte.R
\name{state_fte}
\alias{state_fte}
\title{Calculate State FTE}
\usage{
state_fte(data, yrq)
}
\arguments{
\item{data}{The dbo_class table saved as an R object}

\item{yrq}{The desired YRQ.}
}
\value{
returns a tibble.
}
\description{
Calcualate the State FTE for a given YRQ.
}
\examples{
\dontrun{

dbo_class \%>\%
  state_fte("C012")

}


}
